third_party/blink/web_tests/platform/generic/fast/scrolling/scroll-clears-fragment-anchor-expected.txt

Tests that after an explicit scroll, we do not continue to maintain scroll position at the anchor identified by the hash fragment.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS 10 is 10
PASS successfullyParsed is true

TEST COMPLETE
